Extend bitmasks to 64-bit such that we can support up to 64 CPU cores.
tested by dlg@, ok jsing@

+++ sys/arch/amd64/include/pmap.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: pmap.h,v 1.35 2010/10/26 05:49:10 guenther Exp $	*/
+/*	$OpenBSD: pmap.h,v 1.36 2010/11/30 19:30:16 kettenis Exp $	*/
 /*	$NetBSD: pmap.h,v 1.1 2003/04/26 18:39:46 fvdl Exp $	*/
 
 /*
@@ -318,7 +318,7 @@ struct pmap {
 					/* pointer to a PTP in our pmap */
 	struct pmap_statistics pm_stats;  /* pmap stats (lck by object lock) */
 
-	u_int32_t pm_cpus;		/* mask of CPUs using pmap */
+	u_int64_t pm_cpus;		/* mask of CPUs using pmap */
 };
 
 /*
